# change defer unlock to immediatly unlock to avoid delay call which can speed up all func call